requirements.txt
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
yarl==1.7.2
